requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
langchain>=0.1.0 